feat: persisting architecture data tools
