+++ app/components/dashboard/create-project-dialog.tsx
@@ -1,6 +1,7 @@
 import { useState } from "react";
 import { cn } from "@/lib/utils";
 import { Plus, X } from "lucide-react";
+import * as Dialog from "@radix-ui/react-dialog";
 
 interface CreateProjectDialogProps {
   open: boolean;
@@ -26,105 +27,108 @@ export function CreateProjectDialog({
     setDescription("");
   }
 
-  if (!open) return null;
-
   return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center">
-      {/* Backdrop */}
-      <div
-        className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-        onClick={() => onOpenChange(false)}
-      />
-
-      {/* Dialog */}
-      <div
-        className={cn(
-          "relative z-10 w-full max-w-md rounded-2xl border border-gray-200 bg-white p-6 shadow-2xl",
-          "dark:border-gray-700 dark:bg-gray-900"
-        )}
-      >
-        <div className="flex items-center justify-between">
-          <h2 className="text-xl font-semibold text-gray-900 dark:text-gray-100">
-            New Architecture Project
-          </h2>
-          <button
-            onClick={() => onOpenChange(false)}
-            className="rounded-lg p-1 text-gray-400 transition-colors hover:bg-gray-100 hover:text-gray-600 dark:hover:bg-gray-800"
-          >
-            <X className="h-5 w-5" />
-          </button>
-        </div>
-
-        <form onSubmit={handleSubmit} className="mt-5 space-y-4">
-          <div>
-            <label
-              htmlFor="project-name"
-              className="mb-1.5 block text-sm font-medium text-gray-700 dark:text-gray-300"
-            >
-              Project Name <span className="text-red-500">*</span>
-            </label>
-            <input
-              id="project-name"
-              type="text"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-              placeholder="e.g. E-Commerce Platform"
-              required
-              className={cn(
-                "w-full rounded-lg border border-gray-300 bg-white px-3.5 py-2.5 text-sm",
-                "placeholder:text-gray-400 focus:border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-500/20",
-                "dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100 dark:placeholder:text-gray-500"
-              )}
-            />
+    <Dialog.Root open={open} onOpenChange={onOpenChange}>
+      <Dialog.Portal>
+        <Dialog.Overlay className="fixed inset-0 z-50 bg-black/50 backdrop-blur-sm data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0" />
+        <Dialog.Content
+          className={cn(
+            "fixed left-1/2 top-1/2 z-50 w-full max-w-md -translate-x-1/2 -translate-y-1/2",
+            "rounded-2xl border border-gray-200 bg-white p-6 shadow-2xl",
+            "dark:border-gray-700 dark:bg-gray-900",
+            "data-[state=open]:animate-in data-[state=closed]:animate-out",
+            "data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0",
+            "data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95"
+          )}
+        >
+          <div className="flex items-center justify-between">
+            <Dialog.Title className="text-xl font-semibold text-gray-900 dark:text-gray-100">
+              New Architecture Project
+            </Dialog.Title>
+            <Dialog.Close asChild>
+              <button
+                className="rounded-lg p-1 text-gray-400 transition-colors hover:bg-gray-100 hover:text-gray-600 dark:hover:bg-gray-800"
+                aria-label="Close"
+              >
+                <X className="h-5 w-5" />
+              </button>
+            </Dialog.Close>
           </div>
 
-          <div>
-            <label
-              htmlFor="project-description"
-              className="mb-1.5 block text-sm font-medium text-gray-700 dark:text-gray-300"
-            >
-              Description
-            </label>
-            <textarea
-              id="project-description"
-              value={description}
-              onChange={(e) => setDescription(e.target.value)}
-              placeholder="Brief description of the system's business requirements..."
-              rows={3}
-              className={cn(
-                "w-full resize-none rounded-lg border border-gray-300 bg-white px-3.5 py-2.5 text-sm",
-                "placeholder:text-gray-400 focus:border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-500/20",
-                "dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100 dark:placeholder:text-gray-500"
-              )}
-            />
-          </div>
+          <form onSubmit={handleSubmit} className="mt-5 space-y-4">
+            <div>
+              <label
+                htmlFor="project-name"
+                className="mb-1.5 block text-sm font-medium text-gray-700 dark:text-gray-300"
+              >
+                Project Name <span className="text-red-500">*</span>
+              </label>
+              <input
+                id="project-name"
+                type="text"
+                value={name}
+                onChange={(e) => setName(e.target.value)}
+                placeholder="e.g. E-Commerce Platform"
+                required
+                className={cn(
+                  "w-full rounded-lg border border-gray-300 bg-white px-3.5 py-2.5 text-sm",
+                  "placeholder:text-gray-400 focus:border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-500/20",
+                  "dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100 dark:placeholder:text-gray-500"
+                )}
+              />
+            </div>
 
-          <div className="flex items-center justify-end gap-3 pt-2">
-            <button
-              type="button"
-              onClick={() => onOpenChange(false)}
-              className={cn(
-                "rounded-lg px-4 py-2.5 text-sm font-medium text-gray-700",
-                "transition-colors hover:bg-gray-100",
-                "dark:text-gray-300 dark:hover:bg-gray-800"
-              )}
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              disabled={!name.trim() || isSubmitting}
-              className={cn(
-                "inline-flex items-center gap-2 rounded-lg bg-blue-600 px-4 py-2.5 text-sm font-medium text-white",
-                "transition-colors hover:bg-blue-700 disabled:opacity-50 disabled:cursor-not-allowed"
-              )}
-            >
-              <Plus className="h-4 w-4" />
-              {isSubmitting ? "Creating..." : "Create Project"}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
+            <div>
+              <Dialog.Description asChild>
+                <label
+                  htmlFor="project-description"
+                  className="mb-1.5 block text-sm font-medium text-gray-700 dark:text-gray-300"
+                >
+                  Description
+                </label>
+              </Dialog.Description>
+              <textarea
+                id="project-description"
+                value={description}
+                onChange={(e) => setDescription(e.target.value)}
+                placeholder="Brief description of the system's business requirements..."
+                rows={3}
+                className={cn(
+                  "w-full resize-none rounded-lg border border-gray-300 bg-white px-3.5 py-2.5 text-sm",
+                  "placeholder:text-gray-400 focus:border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-500/20",
+                  "dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100 dark:placeholder:text-gray-500"
+                )}
+              />
+            </div>
+
+            <div className="flex items-center justify-end gap-3 pt-2">
+              <Dialog.Close asChild>
+                <button
+                  type="button"
+                  className={cn(
+                    "rounded-lg px-4 py-2.5 text-sm font-medium text-gray-700",
+                    "transition-colors hover:bg-gray-100",
+                    "dark:text-gray-300 dark:hover:bg-gray-800"
+                  )}
+                >
+                  Cancel
+                </button>
+              </Dialog.Close>
+              <button
+                type="submit"
+                disabled={!name.trim() || isSubmitting}
+                className={cn(
+                  "inline-flex items-center gap-2 rounded-lg bg-blue-600 px-4 py-2.5 text-sm font-medium text-white",
+                  "transition-colors hover:bg-blue-700 disabled:opacity-50 disabled:cursor-not-allowed"
+                )}
+              >
+                <Plus className="h-4 w-4" />
+                {isSubmitting ? "Creating..." : "Create Project"}
+              </button>
+            </div>
+          </form>
+        </Dialog.Content>
+      </Dialog.Portal>
+    </Dialog.Root>
   );
 }
